fix(processors): make registryInt compile on 32-bit platforms (#169)
Comparing a uint against math.MaxInt64 is a compile error when uint is
32 bits (e.g. linux/arm), which broke the GoReleaser build for the
linux_arm_7 target. Widen to uint64 before the comparison.

+++ internal/processors/configuration.go
@@ -427,7 +427,7 @@ func registryInt(value interface{}) (int64, error) {
 	case int64:
 		return v, nil
 	case uint:
-		if v > math.MaxInt64 {
+		if uint64(v) > math.MaxInt64 {
 			return 0, fmt.Errorf("value %d is too large for a registry integer", v)
 		}
 		return int64(v), nil
